Add Entine Solution
-Full solution added
-Functionality to see string length
-Currently no support for text box rows/lines

Program needs a widths.yml file to run. This can be placed in the same directory as the program executable. This yml file can be acquired from the Fonts directory in a CoilSnake project.

Side Note: 132 pixels seems to be the limit for a single line of text before it is carried over to the next line by EarthBound's routine to automatically insert line breaks.